fix: add `onForgotPasswordConfirm` prop - also rename forgot-password-view files
