+++ src/components/forgot-password-view/forgot-password-view.module.scss
